refactor: Change back to FastScanner
Remove FastScanner logic from Ui class
